tests/wpt/tests/websockets/interfaces/WebSocket/events/006.html

<!doctype html>
<title>WebSockets: 'on*' in ws</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.sub.js></script>
<meta name="variant" content="?default">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
test(function(){
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/');
  assert_equals('onopen' in ws, true, 'onopen');
  assert_equals('onmessage' in ws, true, 'onmessage');
  assert_equals('onerror' in ws, true, 'onerror');
  assert_equals('onclose' in ws, true, 'onclose');
});
</script>
